# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi Swagger Documentation for Tasks",
        "description": "L5 Api Rest for educational purposes",
        "version": "0.0.1"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Show task list",
                "description": "Display a listing of the resource.",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "Show all tasks"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            },
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Create a task",
                "description": "Store a newly created resource in storage.",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "content": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Tarea 53",
                                    "description": "Tarea automatica",
                                    "content": "Esta es una tarea de prueba"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task created successfully"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Show task info",
                "description": "Display the specified resource.",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter to search the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Enter a task id number",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show task info"
                    },
                    "404": {
                        "description": "The task has not been found"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            },
            "put": {
                "tags": [
                    "tasks"
                ],
                "summary": "Update a task",
                "description": "Update the specified resource in storage.",
                "operationId": "1aaf37e74360ef4c10fadaee02ae77fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Enter a task id number",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "content": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Tarea 53",
                                    "description": "Tarea automatica",
                                    "content": "Esta es una tarea de prueba"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated success"
                    },
                    "404": {
                        "description": "The task has not been found"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            },
            "delete": {
                "tags": [
                    "tasks"
                ],
                "summary": "Delete a task",
                "description": "Remove the specified resource from storage.",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Enter a task id number",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task delete successfully"
                    },
                    "404": {
                        "description": "The task has been not found"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            }
        }
    }
}